fix:newpページのlogic,viewわけ
